Parameter Addition and Doc Updates
The focus of these changes is to add an additional parameter to the
`New-TestResources.ps1` script to allow more flexibility when using it
to create resources for local testing, where automatically cleaning
up may not be desired.

Additionally, documentation was updated to reflect some of the enhancements
that have been added recently.